Move standalone utilities to their own file.
This is part of the work for #1187.
